pyproject.toml
[project]
name = "kpmg-a2a-audit-compliance"
version = "0.1.0"
description = "Agent-to-Agent Architecture using Google's A2A Framework and LangChain AgentExecutor"
requires-python = ">=3.11"
dependencies = [
    # Core Dependencies
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    
    # LLM & Orchestration
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.1.0",
    
    # Google A2A Framework
    "a2a-sdk>=0.2.5",
    
    # Observability
    "langfuse>=3.0.0",
    
    # OPA Client
    "requests>=2.30.0",
    
    # UI
    "streamlit>=1.30.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "faker>=20.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]